voice-examples/agent/web/openai/backend-python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0

